Minor tweaks
(What I hope to do is eliminate the use of memio.c/h altogether and instead use SDL_RWops in mus2mid.c, as I do elsewhere in the code. Can't firgure out how to handle lines 228-244 in i_music.c yet though.)

+++ src/i_music.c
@@ -241,7 +241,7 @@ void *I_RegisterSong(void *data, int size)
                     mid = malloc(midlen);
                     memcpy(mid, outbuf, midlen);
                     data = mid;
-                    size = midlen;
+                    size = (int)midlen;
                 }
 
                 mem_fclose(instream);
